The mineStats is moved to the MineManager so it can also be used within the Mine code too.
